+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/DI.kt
@@ -373,6 +373,7 @@ fun composeViewModule(messengerConfiguration: MatrixMessengerConfiguration?): Mo
         timelineViewModule(messengerConfiguration),
         userSearchViewModule(),
         settingsViewModule(),
+        appIconPainterModule(),
         verificationViewModule(),
         uiaViewModule(),
         getPlatformNotificationIconModule(),